@@ -11,53 +11,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.3",
-    "@angular/common": "~11.0.3",
-    "@angular/compiler": "~11.0.3",
-    "@angular/core": "~11.0.3",
-    "@angular/forms": "~11.0.3",
-    "@angular/localize": "~11.0.3",
-    "@angular/platform-browser": "~11.0.3",
-    "@angular/platform-browser-dynamic": "~11.0.3",
-    "@angular/router": "~11.0.3",
+    "@angular/animations": "~11.2.2",
+    "@angular/common": "~11.2.2",
+    "@angular/compiler": "~11.2.2",
+    "@angular/core": "~11.2.2",
+    "@angular/forms": "~11.2.2",
+    "@angular/localize": "~11.2.2",
+    "@angular/platform-browser": "~11.2.2",
+    "@angular/platform-browser-dynamic": "~11.2.2",
+    "@angular/router": "~11.2.2",
     "@ctrl/ngx-codemirror": "^4.1.0",
     "@ng-bootstrap/ng-bootstrap": "^9.0.2",
-    "angular-gridster2": "^11.0.0",
+    "angular-gridster2": "^11.1.3",
     "bootstrap": "~4.5.3",
     "bootstrap-darkmode": "^0.7.0",
-    "bootstrap-icons": "^1.3.0",
-    "codemirror": "^5.58.3",
+    "bootstrap-icons": "^1.4.0",
+    "codemirror": "^5.59.4",
     "file-saver": "^2.0.5",
     "keycloak-angular": "^8.1.0",
     "keycloak-js": "12.0.3",
     "ng-bootstrap-darkmode": "^0.4.0",
     "ng2-dragula": "^2.1.1",
     "ngx-clipboard": "^14.0.1",
-    "rxjs": "~6.5.5",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "rxjs": "~6.6.3",
+    "tslib": "^2.1.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.3",
-    "@angular/cli": "~11.0.3",
-    "@angular/compiler-cli": "~11.0.3",
-    "@angular/language-service": "~11.0.3",
-    "@types/codemirror": "^0.0.102",
+    "@angular-devkit/build-angular": "~0.1102.1",
+    "@angular/cli": "~11.2.1",
+    "@angular/compiler-cli": "~11.2.2",
+    "@angular/language-service": "~11.2.2",
+    "@types/codemirror": "^0.0.108",
     "@types/file-saver": "^2.0.1",
-    "@types/jasmine": "~3.6.0",
+    "@types/jasmine": "~3.6.4",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~13.7.7",
-    "codelyzer": "^6.0.0",
+    "@types/node": "~14.14.31",
+    "codelyzer": "^6.0.1",
     "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "jasmine-spec-reporter": "~5.0.2",
+    "karma": "~6.1.1",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.1",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.5"
+    "tslint": "~6.1.3",
+    "typescript": "~4.0.7"
   }
 }
