@@ -1,17 +1,5 @@
 import {Rule} from 'codemirror';
 
-interface Meta {
-  dontIndentStates?: string[];
-  lineComment?: string;
-}
-
-interface SimpleMode {
-  [id: string]: Rule[] | Meta;
-
-  start: Rule[];
-  meta: Meta;
-}
-
 // keywords for fulibScenarios v1.7.0
 export const KEYWORDS = [
   'a',
@@ -121,7 +109,7 @@ export const TYPES = [
   'integer',
 ];
 
-export const SCENARIO_MODE: SimpleMode = {
+export const SCENARIO_MODE: Record<string, Rule[]> & { start: Rule[] } = {
   start: [
     {regex: /(\s*)(##)(.*)$/, token: [null!, 'header', 'comment'], sol: true},
     {regex: /(\s*)(#)(.*)$/, token: [null!, 'header', 'def'], sol: true},
@@ -172,5 +160,5 @@ export const SCENARIO_MODE: SimpleMode = {
   ],
   meta: {
     lineComment: '//',
-  },
+  } as any,
 };
